+++ docs/src/server/markdown.ts
@@ -5,7 +5,7 @@ import * as fs from 'node:fs'
 import * as path from 'node:path'
 import { codeToHtml } from 'shiki'
 import { IGNORE_SYMBOLS, MDN_SYMBOLS } from '../generate/symbols.ts'
-import { routes } from './routes.ts'
+import { getDocsRouteHref, routes } from './routes.ts'
 
 // No types exist for the `frontmatter` package
 const parseFrontmatter = frontmatter.default as unknown as (md: string) => {
@@ -253,23 +253,6 @@ function getShikiExtension(
       ],
     }
   }
-}
-
-function getDocsRouteHref(href: string, version: string | undefined): string | undefined {
-  if (!href.startsWith('/api/')) return undefined
-
-  let url = new URL(href, 'http://localhost')
-  let slug = url.pathname.slice('/api/'.length)
-  if (slug.length === 0) return undefined
-
-  let routeHref: string
-  if (slug.endsWith('.md')) {
-    routeHref = routes.markdown.href({ version, slug: slug.slice(0, -'.md'.length) })
-  } else {
-    routeHref = routes.docs.href({ version, slug: slug.replace(/\/$/, '') })
-  }
-
-  return `${routeHref}${url.search}${url.hash}`
 }
 
 function escapeHtml(value: string): string {
