+++ docs/src/server/router.test.tsx
@@ -4,6 +4,7 @@ import { after, describe, it } from 'remix/test'
 import { assetServer } from './asset-server.ts'
 import { getVersionedLookupHref } from './lookup.ts'
 import { createRouter } from './router.tsx'
+import { getDocsRouteHref } from './routes.ts'
 
 after(async () => {
   await assetServer.close()
@@ -45,6 +46,19 @@ describe('getVersionedLookupHref()', () => {
     assert.equal(
       getVersionedLookupHref('https://example.com/Accept', 'v1.2.3'),
       'https://example.com/Accept',
+    )
+  })
+})
+
+describe('getDocsRouteHref()', () => {
+  it('returns undefined for non-API hrefs', () => {
+    assert.equal(getDocsRouteHref('https://example.com/Accept', 'v1.2.3'), undefined)
+  })
+
+  it('preserves unversioned docs hrefs when version is undefined', () => {
+    assert.equal(
+      getDocsRouteHref('/api/remix/headers/accept/class/Accept.md?tab=docs#example', undefined),
+      '/api/remix/headers/accept/class/Accept.md?tab=docs#example',
     )
   })
 })
