+++ docs/src/server/routes.ts
@@ -7,3 +7,20 @@ export const routes = route({
   lookup: '/(:version/)api.json',
   markdown: '/(:version/)api/*slug.md',
 })
+
+export function getDocsRouteHref(href: string, version: string | undefined): string | undefined {
+  if (!href.startsWith('/api/')) return undefined
+
+  let url = new URL(href, 'http://localhost')
+  let slug = url.pathname.slice('/api/'.length)
+  if (slug.length === 0) return undefined
+
+  let routeHref: string
+  if (slug.endsWith('.md')) {
+    routeHref = routes.markdown.href({ version, slug: slug.slice(0, -'.md'.length) })
+  } else {
+    routeHref = routes.docs.href({ version, slug: slug.replace(/\/$/, '') })
+  }
+
+  return `${routeHref}${url.search}${url.hash}`
+}
